browser-tests: move logon to puppy
